+++ inst/tinytest/test-lincom.R
@@ -7,10 +7,21 @@ dat$cyl <- factor(dat$cyl)
 mod <- lm(mpg ~ carb + cyl, dat)
 
 
-# errors
+# informative errors and warnings
 expect_error(
-    marginaleffects(mod, lincom = "pairwise"),
+    marginaleffects(mod, hypothesis = "pairwise"),
     pattern = "smaller")
+
+expect_warning(
+    marginaleffects(mod, lincom = "pairwise"),
+    pattern = "lincom")
+
+tmp <- lm(mpg ~ wt + drat, data = mtcars)
+expect_error(predictions(
+    mod,
+    newdata = datagrid(wt = 2:3),
+    hypothesis = "wt = drat"),
+    pattern = "unique row")
 
 tmp <- mtcars
 tmp$gear <- factor(tmp$gear)
@@ -19,21 +30,21 @@ expect_error(
     lm(mpg ~ gear, tmp),
     newdata = "mean",
     variables = list(gear = "all"),
-    lincom = "gear = 0"),
-    pattern = "duplicate term names")
+    hypothesis = "gear = 0"),
+    pattern = "row indices")
 
 expect_error(
-    marginaleffects(mod, lincom = "reference"),
+    marginaleffects(mod, hypothesis = "reference"),
     pattern = "smaller")
 
 expect_false(expect_error(
-    marginalmeans(mod, lincom = "pairwise"),
+    marginalmeans(mod, hypothesis = "pairwise"),
     pattern = "smaller"))
 
 expect_error(marginaleffects(
     mod,
     newdata = "mean",
-    lincom = c(1, 1, 1),
+    hypothesis = c(1, 1, 1),
     variables = "cyl"),
     pattern = "of length")
 
@@ -41,21 +52,21 @@ expect_error(marginaleffects(
 expect_error(marginaleffects(
     mod,
     newdata = "mean",
-    lincom = matrix(rep(1, 6), ncol = 2),
+    hypothesis = matrix(rep(1, 6), ncol = 2),
     variables = "cyl"),
     pattern = "2 rows")
 
-# marginaleffects: lincom
+# marginaleffects: hypothesis
 mfx <- marginaleffects(
     mod,
     newdata = "mean",
     variables = "cyl",
-    lincom = "pairwise")
+    hypothesis = "pairwise")
 expect_inherits(mfx, "marginaleffects")
 expect_equivalent(nrow(mfx), 1)
 
 
-# contrasts: lincom
+# contrasts: hypothesis
 cmp1 <- comparisons(
     mod,
     variables = "cyl",
@@ -64,25 +75,25 @@ cmp2 <- comparisons(
     mod,
     variables = "cyl",
     newdata = "mean",
-    lincom = "pairwise")
+    hypothesis = "pairwise")
 expect_equivalent(diff(cmp1$comparison), cmp2$comparison)
 
 
-# marginaleffects: lincom
+# marginaleffects: hypothesis
 mfx <- marginaleffects(
     mod,
     newdata = "mean",
     variables = "cyl",
-    lincom = "pairwise")
+    hypothesis = "pairwise")
 expect_inherits(mfx, "marginaleffects")
 expect_equivalent(nrow(mfx), 1)
 
 
-# predictions: lincom
+# predictions: hypothesis
 p1 <- predictions(
     mod,
     datagrid(cyl = c(4, 6)),
-    lincom = c(-1, 1))
+    hypothesis = c(-1, 1))
 p2 <- predictions(
     mod,
     datagrid(cyl = c(4, 6)))
@@ -95,29 +106,29 @@ lc <- matrix(c(
 p3 <- predictions(
     mod,
     datagrid(cyl = c(4, 6)),
-    lincom = lc)
+    hypothesis = lc)
 expect_inherits(p3, "predictions")
 
 
-# marginalmeans: lincom complex
+# marginalmeans: hypothesis complex
 lc <- c(-2, 1, 1, 0, -1, 1)
 em <- emmeans(mod, "carb") 
 em <- contrast(em, method = data.frame(custom_contrast = lc))
 em <- data.frame(em)
-mm <- marginalmeans(mod, variables = "carb", lincom = lc)
+mm <- marginalmeans(mod, variables = "carb", hypothesis = lc)
 expect_equivalent(mm$marginalmean, em$estimate)
 expect_equivalent(mm$std.error, em$SE)
 
-# marginalmeans: lincom shortcut
-mm <- marginalmeans(mod, variables = "carb", lincom = "reference")
+# marginalmeans: hypothesis shortcut
+mm <- marginalmeans(mod, variables = "carb", hypothesis = "reference")
 expect_equivalent(nrow(mm), 5)
 
-# marginalmeans: lincom complex matrix
+# marginalmeans: hypothesis complex matrix
 lc <- matrix(c(
     -2, 1, 1, 0, -1, 1,
     -1, 1, 0, 0, 0, 0
     ), ncol = 2)
-mm <- marginalmeans(mod, variables = "carb", lincom = lc)
+mm <- marginalmeans(mod, variables = "carb", hypothesis = lc)
 expect_inherits(mm, "marginalmeans")
 expect_equal(nrow(mm), 2)
 
@@ -125,7 +136,7 @@ expect_equal(nrow(mm), 2)
 # marginalmeans: string function
 mm1 <- marginalmeans(
     mod,
-    lincom = "r1 + r2 = 12")
+    hypothesis = "r1 + r2 = 12")
 mm2 <- marginalmeans(mod)
 expect_equivalent(
     mm2$marginalmean[1] + mm2$marginalmean[2] - 12,
@@ -137,11 +148,11 @@ mod <- lm(mpg ~ hp + drat, data = mtcars)
 mfx1 <- marginaleffects(
     mod,
     newdata = "mean",
-    lincom = "exp(r1 + r2) = 100")
+    hypothesis = "exp(r1 + r2) = 100")
 mfx2 <- marginaleffects(
     mod,
     newdata = "mean",
-    lincom = "exp(hp + drat) = 100")
+    hypothesis = "exp(hp + drat) = 100")
 expect_inherits(mfx1, "marginaleffects")
 expect_equivalent(mfx1$dydx, mfx2$dydx)
 expect_equivalent(mfx1$std.error, mfx2$std.error)
@@ -153,11 +164,11 @@ p1 <- predictions(
     newdata = datagrid(hp = c(100, 110, 120)))
 p2 <- predictions(
     mod,
-    lincom = "r1 + r2 + r3 = 10",
+    hypothesis = "r1 + r2 + r3 = 10",
     newdata = datagrid(hp = c(100, 110, 120)))
 p3 <- predictions(
     mod,
-    lincom = "r1 = r2",
+    hypothesis = "r1 = r2",
     newdata = datagrid(hp = c(100, 110, 120)))
 expect_equivalent(sum(p1$predicted) - 10, p2$predicted)
 expect_equivalent(p1$predicted[1] - p1$predicted[2], p3$predicted)